# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Eclypsium BE Challenge API
  version: "1.0.0"
  license:
    name: Proprietary (Take-home submission)
    url: https://github.com/gustavorohrer/asset-manager
  description: |
    REST API for managing assets, components, vulnerabilities, and threats.

    Notes:
    - Unknown query parameters are ignored.
    - Date filters accept RFC3339 datetimes and are compared by date (database columns are DATE).
    - For `sortBy=lastScan`, `NULLS LAST` is applied.
servers:
  - url: https://asset-manager-production-ddd8.up.railway.app
    description: Production (Railway)
security: []
tags:
  - name: Health
    description: Service health endpoint.
  - name: Assets
    description: Asset summary, listing, details, vulnerabilities, threats, update, and delete operations.
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: up
        "500":
          $ref: "#/components/responses/InternalErrorResponse"

  /assets:
    get:
      tags: [Assets]
      summary: List assets
      description: Lists assets with filters, sorting, pagination, and computed vulnerability/threat flags based on findings detected in each component's latest scan.
      operationId: listAssets
      parameters:
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/createdFrom"
        - $ref: "#/components/parameters/createdTo"
        - $ref: "#/components/parameters/lastScanFrom"
        - $ref: "#/components/parameters/lastScanTo"
        - $ref: "#/components/parameters/hasVulnerabilities"
        - $ref: "#/components/parameters/hasThreats"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortOrder"
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAssetsResponse"
              examples:
                success:
                  value:
                    data:
                      - id: AST-001
                        name: Dell PowerEdge R740 Server
                        description: Production database server in datacenter rack A3
                        createdAt: "2024-01-15T00:00:00Z"
                        lastScan: "2024-10-08T00:00:00Z"
                        hasVulnerabilities: true
                        hasThreats: true
                    pagination:
                      page: 1
                      pageSize: 20
                      total: 12
                      totalPages: 1
        "400":
          $ref: "#/components/responses/InvalidQueryParamResponse"
        "500":
          $ref: "#/components/responses/InternalErrorResponse"

  /assets/summary:
    get:
      tags: [Assets]
      summary: Get asset risk summary
      description: Returns the global inventory counters for total assets and assets with findings detected in their latest component scans.
      operationId: getAssetSummary
      responses:
        "200":
          description: Asset summary counters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetRiskSummary"
              example:
                total: 41
                withVulnerabilities: 30
                withThreats: 28
        "500":
          $ref: "#/components/responses/InternalErrorResponse"

  /assets/{id}:
    get:
      tags: [Assets]
      summary: Get asset details
      operationId: getAssetDetails
      parameters:
        - $ref: "#/components/parameters/assetId"
      responses:
        "200":
          description: Asset details with components
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDetailsEnvelope"
              examples:
                success:
                  value:
                    data:
                      id: AST-001
                      name: Dell PowerEdge R740 Server
                      description: Production database server in datacenter rack A3
                      createdAt: "2024-01-15T00:00:00Z"
                      lastScan: "2024-10-08T00:00:00Z"
                      hasVulnerabilities: true
                      hasThreats: true
                      components:
                        - id: CMP-001
                          name: Dell UEFI BIOS
                          version: 2.10.2
                          vendor: Dell Inc.
                          type: UEFI
                          createdAt: "2024-01-15T00:00:00Z"
                          lastScan: "2024-10-08T00:00:00Z"
                          assetId: AST-001
        "400":
          $ref: "#/components/responses/InvalidPathParamResponse"
        "404":
          $ref: "#/components/responses/AssetNotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalErrorResponse"
    patch:
      tags: [Assets]
      summary: Partially update asset properties
      operationId: patchAsset
      parameters:
        - $ref: "#/components/parameters/assetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetRequest"
            examples:
              updateNameDescription:
                value:
                  name: AST-001 Updated
                  description: updated description
              updateLastScan:
                value:
                  lastScan: "2024-10-07T00:00:00Z"
              clearLastScan:
                value:
                  lastScan: null
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetUpdatedEnvelope"
              examples:
                success:
                  value:
                    data:
                      id: AST-001
                      name: AST-001 Updated
                      description: updated description
                      createdAt: "2024-01-15T00:00:00Z"
                      lastScan: "2024-10-07T00:00:00Z"
        "400":
          $ref: "#/components/responses/InvalidRequestBodyResponse"
        "404":
          $ref: "#/components/responses/AssetNotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalErrorResponse"
    delete:
      tags: [Assets]
      summary: Delete asset (hard delete)
      operationId: deleteAsset
      parameters:
        - $ref: "#/components/parameters/assetId"
      responses:
        "200":
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDeletedEnvelope"
              example:
                data:
                  id: AST-001
                  deleted: true
        "400":
          $ref: "#/components/responses/InvalidPathParamResponse"
        "404":
          $ref: "#/components/responses/AssetNotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalErrorResponse"

  /assets/{id}/vulnerabilities:
    get:
      tags: [Assets]
      summary: List vulnerabilities by asset
      operationId: listAssetVulnerabilities
      parameters:
        - $ref: "#/components/parameters/assetId"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/severity"
      responses:
        "200":
          description: Vulnerabilities from latest scan per component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAssetVulnerabilitiesResponse"
              examples:
                success:
                  value:
                    data:
                      - id: VUL-001
                        description: Dell UEFI BIOS vulnerable to BootHole
                        severity: CRITICAL
                        scanId: SCN-001
                        componentId: CMP-001
                        componentName: Dell UEFI BIOS
                        performedAt: "2024-10-08T00:00:00Z"
                    pagination:
                      page: 1
                      pageSize: 10
                      total: 1
                      totalPages: 1
        "400":
          $ref: "#/components/responses/InvalidQueryParamResponse"
        "404":
          $ref: "#/components/responses/AssetNotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalErrorResponse"

  /assets/{id}/threats:
    get:
      tags: [Assets]
      summary: List threats by asset
      operationId: listAssetThreats
      parameters:
        - $ref: "#/components/parameters/assetId"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/riskLevel"
      responses:
        "200":
          description: Threats from latest scan per component
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAssetThreatsResponse"
              examples:
                success:
                  value:
                    data:
                      - id: THR-001
                        description: Bootkits can bypass SecureBoot
                        riskLevel: HIGH
                        type: Firmware Implant
                        scanId: SCN-001
                        componentId: CMP-001
                        componentName: Dell UEFI BIOS
                        performedAt: "2024-10-08T00:00:00Z"
                    pagination:
                      page: 1
                      pageSize: 10
                      total: 1
                      totalPages: 1
        "400":
          $ref: "#/components/responses/InvalidQueryParamResponse"
        "404":
          $ref: "#/components/responses/AssetNotFoundResponse"
        "500":
          $ref: "#/components/responses/InternalErrorResponse"

components:
  parameters:
    assetId:
      name: id
      in: path
      required: true
      description: Asset identifier
      schema:
        type: string
      example: AST-001
    name:
      name: name
      in: query
      required: false
      description: Contains filter (case-insensitive)
      schema:
        type: string
      example: router
    createdFrom:
      name: created_from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      example: "2024-01-01T00:00:00Z"
    createdTo:
      name: created_to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      example: "2024-12-31T23:59:59Z"
    lastScanFrom:
      name: last_scan_from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      example: "2024-09-01T00:00:00Z"
    lastScanTo:
      name: last_scan_to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      example: "2024-10-31T23:59:59Z"
    hasVulnerabilities:
      name: has_vulnerabilities
      in: query
      required: false
      description: Filters assets by whether their latest component scans contain vulnerabilities.
      schema:
        type: boolean
      example: true
    hasThreats:
      name: has_threats
      in: query
      required: false
      description: Filters assets by whether their latest component scans contain threats.
      schema:
        type: boolean
      example: false
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 1
      example: 1
    pageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    sortBy:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        enum: [createdAt, name, lastScan]
        default: createdAt
      example: createdAt
    sortOrder:
      name: sortOrder
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      example: desc
    severity:
      name: severity
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Severity"
      example: HIGH
    riskLevel:
      name: riskLevel
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RiskLevel"
      example: HIGH

  responses:
    InvalidQueryParamResponse:
      description: Invalid query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: INVALID_QUERY_PARAM
              message: one or more query parameters are invalid
              details:
                - field: page
                  issue: must be a positive integer
                  value: "0"
    InvalidPathParamResponse:
      description: Invalid path parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: INVALID_PATH_PARAM
              message: asset id is required
    InvalidRequestBodyResponse:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidBody:
              value:
                error:
                  code: INVALID_REQUEST_BODY
                  message: request body is invalid
                  details:
                    - field: id
                      issue: is not allowed
                      value: "\"AST-002\""
                    - field: body
                      issue: "must include at least one updatable field: name, description, lastScan"
                      value: ""
                    - field: body
                      issue: must include at least one valid updatable field
                      value: ""
            bodyTooLarge:
              value:
                error:
                  code: INVALID_REQUEST_BODY
                  message: request body exceeds maximum allowed size
    AssetNotFoundResponse:
      description: Asset not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: ASSET_NOT_FOUND
              message: asset not found
    InternalErrorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: INTERNAL_ERROR
              message: internal server error

  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [up]

    AssetRiskSummary:
      type: object
      required: [total, withVulnerabilities, withThreats]
      properties:
        total:
          type: integer
        withVulnerabilities:
          type: integer
        withThreats:
          type: integer

    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    Severity:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]

    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH]

    AssetSummary:
      type: object
      required:
        - id
        - name
        - description
        - createdAt
        - hasVulnerabilities
        - hasThreats
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        lastScan:
          type: string
          format: date-time
          nullable: true
        hasVulnerabilities:
          type: boolean
        hasThreats:
          type: boolean

    ListAssetsResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssetSummary"
        pagination:
          $ref: "#/components/schemas/Pagination"

    AssetComponent:
      type: object
      required: [id, name, version, vendor, type, createdAt, assetId]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        vendor:
          type: string
        type:
          type: string
        createdAt:
          type: string
          format: date-time
        lastScan:
          type: string
          format: date-time
          nullable: true
        assetId:
          type: string

    AssetDetails:
      type: object
      required:
        - id
        - name
        - description
        - createdAt
        - hasVulnerabilities
        - hasThreats
        - components
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        lastScan:
          type: string
          format: date-time
          nullable: true
        hasVulnerabilities:
          type: boolean
        hasThreats:
          type: boolean
        components:
          type: array
          items:
            $ref: "#/components/schemas/AssetComponent"

    AssetDetailsEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AssetDetails"

    AssetVulnerability:
      type: object
      required: [id, description, severity, scanId, componentId, componentName, performedAt]
      properties:
        id:
          type: string
        description:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        scanId:
          type: string
        componentId:
          type: string
        componentName:
          type: string
        performedAt:
          type: string
          format: date-time

    ListAssetVulnerabilitiesResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssetVulnerability"
        pagination:
          $ref: "#/components/schemas/Pagination"

    AssetThreat:
      type: object
      required: [id, description, riskLevel, type, scanId, componentId, componentName, performedAt]
      properties:
        id:
          type: string
        description:
          type: string
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        type:
          type: string
        scanId:
          type: string
        componentId:
          type: string
        componentName:
          type: string
        performedAt:
          type: string
          format: date-time

    ListAssetThreatsResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssetThreat"
        pagination:
          $ref: "#/components/schemas/Pagination"

    UpdateAssetRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
          description: Trimmed server-side; must contain at least one non-whitespace character; null is not allowed.
        description:
          type: string
          maxLength: 10000
          description: null is not allowed.
        lastScan:
          type: string
          format: date-time
          nullable: true
          description: RFC3339 datetime, or null to clear.

    AssetUpdated:
      type: object
      required: [id, name, description, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        lastScan:
          type: string
          format: date-time
          nullable: true

    AssetUpdatedEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AssetUpdated"

    AssetDeleted:
      type: object
      required: [id, deleted]
      properties:
        id:
          type: string
        deleted:
          type: boolean

    AssetDeletedEnvelope:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AssetDeleted"

    QueryValidationDetail:
      type: object
      required: [field, issue, value]
      properties:
        field:
          type: string
        issue:
          type: string
        value:
          type: string

    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/QueryValidationDetail"

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ApiError"